rename SchemedTextField to SchemedTextFieldArray
